Implement seat-based messaging feature and improve DTOs
Added the ability for users to send messages to others based on seat allocation. Introduced `MessageSendRequestDto` for message content validation and updated error codes. Included Swagger annotations for API documentation and enhanced user and authentication-related DTOs with schema details. Minor bug fixes and refactoring for consistency.

+++ backend/src/main/java/com/shhtudy/backend/repository/SeatRepository.java
@@ -3,5 +3,5 @@ package com.shhtudy.backend.repository;
 import com.shhtudy.backend.entity.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
 }
